backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
beautifulsoup4==4.12.3
requests==2.31.0
aiohttp==3.9.3
openai==1.12.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.9
pydantic==2.6.1
pytest==8.0.0
httpx==0.26.0
alembic==1.13.1
redis==5.0.1
celery==5.3.6
feedparser==6.0.11
tweepy==4.14.0
pytrends==4.9.3 